Rename component Navigation 2 -> Header
