--- src/components/Navigation2/index.ts
@@ -1,1 +1,0 @@
-export { default as Navigation2 } from './Navigation2';
